pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prosense"
version = "0.1.0"
description = "Multimodal physiological signal processing framework for EEG, PPG, ACC, GYRO, BVP, GSR, and TEMP data"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ProSense Contributors", email = "prosense@example.com"}
]
keywords = [
    "physiological-signals",
    "signal-processing",
    "EEG",
    "PPG",
    "feature-extraction",
    "biosignals",
    "neuroscience",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy==1.26.4",
    "scipy==1.11.4",
    "pandas==2.1.4",
    "mne==1.6.1",
    "matplotlib==3.8.2",
    "seaborn==0.13.1",
    "pyyaml==6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.1",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.2",
]

test = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

lint = [
    "black>=23.12.1",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.2",
]

docs = [
    "sphinx>=7.2.6",
    "sphinx-rtd-theme>=2.0.0",
]

all = [
    "prosense[dev,test,lint,docs]",
]

[project.urls]
Homepage = "https://github.com/siddhant61/ProSense"
Repository = "https://github.com/siddhant61/ProSense"
Issues = "https://github.com/siddhant61/ProSense/issues"

[project.scripts]
prosense = "main:main"

[tool.setuptools]
packages = ["prosense", "prosense.modalities"]
package-dir = {"prosense" = ".", "prosense.modalities" = "modalities"}

[tool.setuptools.package-data]
prosense = ["config.yaml", "*.md"]

# =============================================================================
# PYTEST CONFIGURATION
# =============================================================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",                           # Verbose output
    "--strict-markers",             # Strict marker checking
    "--tb=short",                   # Shorter traceback format
    "--cov=.",                      # Coverage for all project files
    "--cov-report=term-missing",    # Show missing lines in coverage
    "--cov-report=html",            # Generate HTML coverage report
    "--cov-report=xml",             # Generate XML coverage report
    "--cov-fail-under=30",          # Fail if coverage below 30%
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "requires_data: marks tests that require actual data files",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# =============================================================================
# BLACK CODE FORMATTER
# =============================================================================
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.tox
  | build
  | dist
  | __pycache__
  | \.eggs
  | \.pytest_cache
)/
'''

# =============================================================================
# ISORT IMPORT SORTER
# =============================================================================
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [".venv", "venv", ".tox", "build", "dist"]

# =============================================================================
# MYPY TYPE CHECKER
# =============================================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "mne.*",
    "scipy.*",
    "matplotlib.*",
    "seaborn.*",
]
ignore_missing_imports = true

# =============================================================================
# COVERAGE
# =============================================================================
[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "setup.py",
    "log_parser.py",  # Orphaned code
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
